Refactor main loop to not use geometry
In order to find the lowest point inside a
Mesh, it is necessary to traverse the geometry and
perform some calculations.

The geometry of the Mesh is of type
BufferGeometry and it is converted to normal
Geometry in order to perform the calculations.

The problem is that during the conversion from
BufferGeometry to Geometry, data about
skinIndices and skinWeights is lost (either a bug or
it was inteded to behave like this).

The current workaround is to modify the three.js
library such that it preserves the required data
during conversion.

This commit eliminates the need to convert the
geometry from BufferGeometry to normal Geometry
by getting all the data straight from the attributes
of the buffered geometry. It then performs the exact
same calculations in order to find out the lowest point in the geometry.

This closes #15, closes #16