fix transform controls and orbit controls triggered at same time
The version of transform controls I copied has removed
the _event.preventDefault()_ from the body of onMouseHover.
The _preventDefault_ call prevented the event from being captured
by something else (in this case the orbit controls).

In order to not allow orbit controls and transform
to be used at the same time, while transform controls are being used,
the orbitcontrols will be disabled.