refactor(UploadWizard): mount/unmount instead of hide
Refactor Uploadwizard to mount or unmount instead of changing display
to none to hide.
Do the same for all wizard steps (only mount the component for the
current step).
This makes it easier to manage the components for each individual step
by not having to deal with null/undefined props.

BROKEN COMMIT (refactoring continues in following commits)